moved make_star_true into journalist_app/utils.py
